refactor helper and decider
